+++ src/domain/flow-system/src/services/flow/flow_service.rs
@@ -15,13 +15,11 @@ use tokio_stream::Stream;
 
 use crate::{
     DatasetFlowFilters,
-    DatasetFlowType,
     FlowID,
     FlowKey,
     FlowPaginationOpts,
     FlowState,
     SystemFlowFilters,
-    SystemFlowType,
 };
 
 /////////////////////////////////////////////////////////////////////////////////////////
@@ -65,20 +63,6 @@ pub trait FlowService: Sync + Send {
         &self,
         pagination: FlowPaginationOpts,
     ) -> Result<FlowStateListing, ListFlowsError>;
-
-    /// Returns state of the latest flow of certain type created for the given
-    /// dataset
-    async fn get_last_flow_by_dataset_of_type(
-        &self,
-        dataset_id: &DatasetID,
-        flow_type: DatasetFlowType,
-    ) -> Result<Option<FlowState>, GetLastDatasetFlowError>;
-
-    /// Returns state of the latest system flow of certain type
-    async fn get_last_system_flow_of_type(
-        &self,
-        flow_type: SystemFlowType,
-    ) -> Result<Option<FlowState>, GetLastSystemFlowError>;
 
     /// Returns current state of a given flow
     async fn get_flow(&self, flow_id: FlowID) -> Result<FlowState, GetFlowError>;
@@ -198,12 +182,22 @@ impl From<LoadError<FlowState>> for CancelScheduledTasksError {
 
 #[derive(Debug)]
 pub struct FlowServiceRunConfig {
+    /// Defines discretion for main scheduling loop: how often new data is
+    /// checked and processed
     pub awaiting_step: chrono::Duration,
+    /// Defines minimal time between 2 runs of the same flow configuration
+    pub mandatory_throttling_period: chrono::Duration,
 }
 
 impl FlowServiceRunConfig {
-    pub fn new(awaiting_step: chrono::Duration) -> Self {
-        Self { awaiting_step }
+    pub fn new(
+        awaiting_step: chrono::Duration,
+        mandatory_throttling_period: chrono::Duration,
+    ) -> Self {
+        Self {
+            awaiting_step,
+            mandatory_throttling_period,
+        }
     }
 }