+++ src/infra/flow-system-inmem/tests/tests/utils/mod.rs
@@ -7,9 +7,10 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
-mod test_flow_configuration_service_inmem;
-mod test_flow_event_store_inmem;
-mod test_flow_service_inmem;
+mod flow_system_test_listener;
+mod manual_flow_trigger_driver;
+mod task_driver;
 
-mod utils;
-pub(crate) use utils::*;
+pub(crate) use flow_system_test_listener::*;
+pub(crate) use manual_flow_trigger_driver::*;
+pub(crate) use task_driver::*;