+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+ - Flow system scheduling rules improved to respect system-wide throttling setting and 
+    take last succesful run into account when rescheduling a flow or after restart
+
 ## [0.156.2] - 2024-02-07
 ### Changed
 - Using unique container and network names to prevent collisions when running concurrent `kamu` processes